Refactored QueryMode into QueryMode and GeoType. QueryMode indicates whether we're looking for sites or units; GeoType indicates whether we should match on city name or zip code as well as our bounding box around the centroid. Still not perfect, but better. A side effect is that we also got rid of that minInventory() nonsense in SearchCriteria.
Once that was done, I was able to change the query generation (I had to switch from Hibernate criteria to HQL) so we now use the DB to find sites with available inventory, rather than pulling all candidates into memory, and pruning in the app. Now if we can just use the DB to do our sorting, then perhaps we can even have paginated result sets in our UI. Woo-hoo!